+++ src/components/Sections/Resume/index.tsx
@@ -8,13 +8,13 @@ export default function Resume() {
     return (
         <Section className="bg-neutral-100" sectionId={SectionId.Resume}>
             <div className="flex flex-col divide-y-2 divide-neutral-300">
-                <ResumeSection title="Education">
-                    {education.map((item, index) => (
+                <ResumeSection title="Work">
+                    {experience.map((item, index) => (
                         <TimelineItem item={item} key={`${item.title}-${index}`} />
                     ))}
                 </ResumeSection>
-                <ResumeSection title="Work">
-                    {experience.map((item, index) => (
+                <ResumeSection title="Education">
+                    {education.map((item, index) => (
                         <TimelineItem item={item} key={`${item.title}-${index}`} />
                     ))}
                 </ResumeSection>